Feature/health

+++ Assets/Scripts/Player/Aspect.cs
@@ -9,12 +9,12 @@ namespace Player
         public Scrollbar scrollbar;
         public TMP_Text topText;
         public TMP_Text bottomText;
-        public int value = 5;
+        public float value = 5;
         public int maxValue = 10;
 
         public void SetSliderValue(float value)
         {
-            int sliderValue = (int)(value * 10);
+            float sliderValue = (value * 10);
             topText.text = sliderValue.ToString();
             bottomText.text = (maxValue - sliderValue).ToString();
             this.value = sliderValue;
